Create faulty_multiple_linear_regression.py
Why am I uploading a code that I know is faulty? Because results don't always matter. The manner of writing of the code is so beautiful (at least for me, sorry for trying to appreciate my own code!) that I thought I must include it even though the error function graph generated by it is exactly the opposite of what it should be (it is increasing instead of decreasing). Perhaps a faulty approach opposite in logic created output which is opposite than desired. Anyways a good learning experience!